test: 잘못된 클래스명 수정
ThemeE2ETest -> MemberE2ETest
